pyproject.toml
[project]
name = "fastapi-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "Christian Viegas",email = "chris.gabriel346@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.18.0,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "fastapi_zero", from = "src"}]

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT", "FAST"]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/fastapi_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src/fastapi_zero -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]
